% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_probs_dfs}
\alias{add_probs_dfs}
\title{addition of probability-data-frames}
\usage{
add_probs_dfs(PREDS_LST)
}
\arguments{
\item{PREDS_LST}{is a list of data frames}
}
\value{
average of a number of data frames
}
\description{
this function takes a number of probability data frames and returns their average (mainly used in multi-class classification).
}
