% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds_and_data_split.R
\name{class_folds}
\alias{class_folds}
\title{stratified folds (in classification)}
\usage{
class_folds(folds, RESP, shuffle = FALSE)
}
\arguments{
\item{folds}{is an integer specifying the number of folds}

\item{RESP}{is the response variable}

\item{shuffle}{is a boolean specifying if the vector of indices should be shuffled or not}
}
\value{
a list of indices
}
\description{
this function creates stratified folds in binary and multiclass classification
}
\examples{

\dontrun{

data(iris)

y = as.factor(iris[, 5])

folds = class_folds(10, y, shuffle = TRUE)
}
}
