% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds_and_data_split.R
\name{DataSplit}
\alias{DataSplit}
\title{partition of data (train-test-split)}
\usage{
DataSplit(y, TrainRatio = 0.75, regression = TRUE, shuffle = FALSE)
}
\arguments{
\item{y}{is a numeric vector (response variable)}

\item{TrainRatio}{is the percentage of train-data after the partition}

\item{regression}{is a boolean (TRUE, FALSE) indicating if it's a regression or classification task}

\item{shuffle}{is a boolean (TRUE, FALSE) indicating if the data should be shuffled or not (by default 5 times)}
}
\value{
a list of indices (train-test)
}
\description{
partition of data (train-test-split)
}
\examples{

\dontrun{

data(iris)

y = X[, 1]

split = DataSplit(y, TrainRatio = 0.75, regression = FALSE, shuffle = FALSE)

}
}
