% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_of_multiple_predictors.R
\name{func_correlation}
\alias{func_correlation}
\title{find correlated variables}
\usage{
func_correlation(data, target = NULL, correlation_thresh = NULL,
  use_obs = NULL, correlation_method = NULL)
}
\arguments{
\item{data}{either a data frame or a matrix}

\item{target}{either a string (name of the predictor/response in the data set) or a vector of strings (predictor/response names of the data set)}

\item{correlation_thresh}{a float indicating the correlation threshold}

\item{use_obs}{one of "everything", "all.obs", "complete.obs", "na.or.complete", "pairwise.complete.obs"}

\item{correlation_method}{one of "pearson", "kendall", "spearman"}
}
\value{
either a data frame or a list of data frames
}
\description{
This function takes a data frame or a matrix and returns either a data frame or a list of data frames with correlated variables
}
\details{
This function takes a data frame or a matrix and returns the correlated variables of the data
}

