% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_shuffle}
\alias{func_shuffle}
\title{shuffle data}
\usage{
func_shuffle(vec, times = 10)
}
\arguments{
\item{vec}{is a vector of indices}

\item{times}{is a number}
}
\value{
shuffled indices of a vector
}
\description{
this function shuffles the items of a vector
}

