% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalized}
\alias{normalized}
\title{normalize data}
\usage{
normalized(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
normalized data in form of a vector
}
\description{
this function normalizes the feature importance of the algorithms, so that data between algorithms is in the same scale
}
