% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds_and_data_split.R
\name{regr_folds}
\alias{regr_folds}
\title{create folds (in regression)}
\usage{
regr_folds(folds, RESP, stratified = FALSE)
}
\arguments{
\item{folds}{is an integer specifying the number of folds}

\item{RESP}{is the response variable}

\item{stratified}{is a boolean specifying if the folds should be stratfied}
}
\value{
a list of indices
}
\description{
this function creates both stratified and non-stratified folds in regression
}
\examples{

\dontrun{

data(iris)

y = X[, 1]

folds = regr_folds(5, y, stratified = FALSE)

}
}
