% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_of_multiple_predictors.R
\name{remove_duplic_func}
\alias{remove_duplic_func}
\title{function to remove duplicated pairs of predictors ( used in the func_correlation )}
\usage{
remove_duplic_func(sublist)
}
\description{
this is a secondary function that is used in the func_correlation
}
\keyword{internal}
