% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_of_multiple_predictors.R
\name{second_func_cor}
\alias{second_func_cor}
\title{secondary function ( used in the func_correlation )}
\usage{
second_func_cor(dat_frame)
}
\description{
this is a secondary function that is used in the func_correlation
}
\keyword{internal}
